+++ src/test/java/chess/domain/BoardInitializerTest.java
@@ -1,6 +1,6 @@
 package chess.domain;
 
-import chess.domain.manager.BoardInitializer;
+import chess.domain.game.BoardInitializer;
 import chess.domain.pieces.*;
 import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
